LayoutTests/http/tests/cache/memory-cache-pruning-expected.txt

This test exercices the code path executed when the memory cache reaches its maximum size

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS previousSize > 0 is true
PASS internals.memoryCacheSize() < previousSize is true
PASS successfullyParsed is true

TEST COMPLETE
